+++ Nos/Views/ThreadView.swift
@@ -128,34 +128,6 @@ struct ThreadView: View {
         }
     }
 }
-
-struct ExpandingTextFieldAndSubmitButton: View {
-    var placeholder: String
-    @Binding var reply: String
-    var action: () -> Void
-    
-    var body: some View {
-        ZStack(alignment: .trailing) {
-            TextField(placeholder, text: $reply, axis: .vertical)
-                .lineLimit(5)
-                .textFieldStyle(.roundedBorder)
-                .padding(.trailing, 30)
-            Button(
-                action: {
-                    self.action()
-                    reply = ""
-                },
-                label: {
-                    Image(systemName: "paperplane.fill")
-                        .font(.headline)
-                }
-            )
-        }
-        .padding(.trailing, 8)
-        .opacity(1.0)
-    }
-}
-
 struct ThreadView_Previews: PreviewProvider {
     static var persistenceController = PersistenceController.preview
     static var previewContext = persistenceController.container.viewContext